+++ app/ui-message/client/messageBox/messageBoxActions.js
@@ -1,8 +1,8 @@
+import mime from 'mime-type/with-db';
 import { Meteor } from 'meteor/meteor';
 import { Random } from 'meteor/random';
 import { ReactiveVar } from 'meteor/reactive-var';
 import { Tracker } from 'meteor/tracker';
-import mime from 'mime-type/with-db';
 import { VRecDialog } from '../../../ui-vrecord/client';
 import { messageBox, modal } from '../../../ui-utils';
 import { fileUpload } from '../../../ui';
