+++ app/ui-message/client/messageBox/messageBoxAutogrow.js
