+++ app/ui-message/client/messageBox/messageBoxFormatting.js
@@ -1,6 +1,6 @@
-import { katex } from '../../katex/client';
-import { Markdown } from '../../markdown/client';
-import { settings } from '../../settings';
+import { katex } from '../../../katex/client';
+import { Markdown } from '../../../markdown/client';
+import { settings } from '../../../settings';
 
 
 export const formattingButtons = [
