+++ app/ui-message/client/messageBox/messageBoxNotSubscribed.js
@@ -1,11 +1,11 @@
 import { Meteor } from 'meteor/meteor';
 import { Session } from 'meteor/session';
 import { Template } from 'meteor/templating';
-import { settings } from '../../settings';
-import { call } from '../../ui-utils';
-import { roomTypes } from '../../utils';
-import { RoomManager, RoomHistoryManager } from '../../ui-utils';
-import { hasAllPermission } from '../../authorization';
+import { settings } from '../../../settings';
+import { call } from '../../../ui-utils';
+import { roomTypes } from '../../../utils';
+import { RoomManager, RoomHistoryManager } from '../../../ui-utils';
+import { hasAllPermission } from '../../../authorization';
 import './messageBoxNotSubscribed.html';
