+++ app/ui-message/client/messageBox/messageBoxTyping.html
