+++ app/ui-message/client/messageBox/messageBoxTyping.js
@@ -1,6 +1,6 @@
 import { Template } from 'meteor/templating';
-import { MsgTyping } from '../../ui';
-import { t } from '../../utils';
+import { MsgTyping } from '../../../ui';
+import { t } from '../../../utils';
 import './messageBoxTyping.html';
