--- app/ui-message/client/startup/index.js
@@ -1,1 +1,0 @@
-import './messageBoxActions';
